requirements.txt
aniso8601==1.2.1
click==6.7
Flask==0.12.2
Flask-HTTPAuth==3.2.3
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
PyJWT==1.4.2
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
